+++ src/services/Authentication/core/Common/Constants/Constants.cs
@@ -1,4 +1,4 @@
-﻿namespace Common;
+﻿namespace Common.Constants;
 
 public static class Constants
 {
