+++ src/services/Authentication/core/Common/Constants/RoleConstants.cs
@@ -1,4 +1,4 @@
-namespace Common;
+namespace Common.Constants;
 
 public static class RoleConstants
 {
