+++ src/services/Authentication/src/Authentication.DataAccessLayer/Entities/ConfigurationModels/JwtConfiguration.cs
@@ -1,4 +1,4 @@
-﻿using Common;
+﻿using Common.Constants;
 
 namespace Authentication.DataAccessLayer.Entities.ConfigurationModels;
