+++ src/services/Authentication/src/Authentication.PresentationLayer/Controllers/AuthenticationController.cs
@@ -1,6 +1,6 @@
 using Authentication.BusinessLogicLayer.DataTransferObjects;
 using Authentication.BusinessLogicLayer.Services.Interfaces;
-using Common;
+using Common.Constants;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
